% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{directors}
\alias{directors}
\title{List of directors}
\format{
This data frame contains the following columns:
\describe{
\item{\code{story_number}}{Story number}
\item{\code{director}}{Episode's director}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Doctor_Who_episodes_(2005-present)}
}
\usage{
directors
}
\description{
List of directors for all episodes from the revived Doctor Who era (2005-present)
}
\examples{
library(dplyr)
directors \%>\%
    count(director) \%>\%
    arrange(desc(n)) \%>\%
    head(5)
}
\keyword{datasets}
