% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{drwho_directors}
\alias{drwho_directors}
\title{List of Dr Who directors}
\format{
This data frame contains the following columns:
\describe{
\item{\code{story_number}}{Story number}
\item{\code{director}}{Episode's director}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Doctor_Who_episodes_(2005-present)}
}
\usage{
drwho_directors
}
\description{
List of directors for all episodes from the revived Doctor Who era (2005-present)
}
\examples{
library(dplyr)
drwho_directors |>
    count(director, sort = TRUE) |>
    head(5)
}
\keyword{datasets}
