% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{drwho_writers}
\alias{drwho_writers}
\title{List of Dr Who writers}
\format{
This data frame contains the following columns:
\describe{
\item{\code{story_number}}{Story number}
\item{\code{writer}}{Episode's writer}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Doctor_Who_episodes_(2005-present)}
}
\usage{
drwho_writers
}
\description{
List of writers for all episodes from the revived Doctor Who era (2005-present)
}
\examples{
library(dplyr)
drwho_writers \%>\%
    count(writer) \%>\%
    arrange(desc(n)) \%>\%
    head(5)
}
\keyword{datasets}
