% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{torchwood_writers}
\alias{torchwood_writers}
\title{List of Torchwood writers}
\format{
This data frame contains the following columns:
\describe{
\item{\code{story_number}}{Story number}
\item{\code{writer}}{Episode's writer}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Torchwood_episodes}
}
\usage{
torchwood_writers
}
\description{
List of directors for all episodes from Torchwood series (2006-2011)
}
\examples{
library(dplyr)
torchwood_writers \%>\%
    count(writer) \%>\%
    arrange(desc(n)) \%>\%
    head(5)
}
\keyword{datasets}
